Fix import-layer violation: move build_user_ssh_command to primitives
The hosts fix put build_user_ssh_command in hosts/common.py, but
interfaces/provider_instance.py (a lower layer) imports it, violating the
mngr layers contract (interfaces must not import hosts). The helper is a pure
string builder for SSHInfo.command, so it belongs in primitives.py next to
SSHInfo -- the lowest layer, importable by interfaces, api, and the provider
projects alike. Relocated the function and its two tests; updated all importers
(provider_instance, api/discovery_events, mngr_modal, mngr_vps_docker).
